ライブキャストの通知処理の為のクラスのリファクタリング
